+++ src/middleware/analyse.js
@@ -1,39 +1,39 @@
-import E from '../error';
-import _ from 'lodash';
+import E from '../error'
+import _ from 'lodash'
 
 export default async (ctx, next) => {
-  let req = ctx.request;
-  let path = req.url;
+  let req = ctx.request
+  let path = req.url
   //如果调用的是api路由，则进行分析和拦截
   if(path === '/api'){
     //验证安全性并进行日志记录
-    let postData = ctx.request.body;
+    let postData = ctx.request.body
     // 从缓存中获取已经授权的应用列表
-    let apps = await ctx.getAppsFromCache();
+    let apps = await ctx.fpm.getClients()
     // 验证应用授权信息
-    let appItem = apps[postData.appkey];
+    let appItem = apps[postData.appkey]
     if(appItem.approot){
-      let approot = appItem.approot;
+      let approot = appItem.approot
       if(approot === '*'){
         // 使用了通配权限
-        await next();
+        await next()
       }else{
-        let method = postData.method;
-        let scope = method.split('.')[0];
-        let roots = approot.split(',');
+        let method = postData.method
+        let scope = method.split('.')[0]
+        let roots = approot.split(',')
         // 涵盖权限
-        if(_.indexOf(roots,scope) > -1){
-          await next();
+        if(_.indexOf(roots, scope) > -1){
+          await next()
         }else{
-          ctx.fail(E.System.ROOT_ERROR);
+          ctx.fail(E.System.ROOT_ERROR)
         }
       }
 
     }else{
-      ctx.fail(E.System.ROOT_ERROR);
+      ctx.fail(E.System.ROOT_ERROR)
     }
 
   }else{
-    await next();
+    await next()
   }
 }